Remove need for manual specification of the last .control-group
Until now we needed to explicitly mark the last .control-group container
so that it would get a border at the bottom.

Because we now style those containers in a different (smarter) way,
there is no longer a need to use the "last" class on the last container.

This makes templates simpler and also makes sure that they're always
styled correctly.