Optimize AngularJS performance
If the services list contained ~70 services, it was getting
unbearably slow, causing 1-3 second freezes in both Firefox and Chrome.

Turns out:

1. $timeout needs to be called with invokeApply = false (3rd argument),
when the callback does not need to trigger a $digest.
Because we didn't use it like that and were using many timeouts,
we were triggering hundreds of digest cycles unintentionally.

2. Filters are called many times and it's better to avoid them
and cache the results on the scope from the directive's code

3. Certain elements don't need to be fully generated (the "Add service" dropdown)
until the dropdown is actually expanded/shown

4. $http with {"cache": true} could be used for certain APIs to avoid
them being hit over and over again

5. Certain APIs don't need to be hit until actually needed (the "Add
service" dropdown using the commands list)

6. Certain expressions could be precaculated to save time (and simplify
templates): $scope.jsTimestamp = ($scope.timestamp * 1000);