Do not allow in-use time periods to be deleted
Not that anything fatal will happen.
The ContactEventsSubscriber will notice the time-period delete operation
and will clean up everything that references it.
That means all contact users will be deleted, which sounds quite bad.

We can't allow a "let's delete this contact period" to have such a
disastrous effect.

An alternative would be to not delete contact that references the
time-period, but to set its time-period to something else.
It's most logical to set it to NULL.
But then we'll have to worry about contacts potentially having
a NULL time-period.. Better not.