Update composer.json and README - PHP 5.4 is required.
According to 08b767281237e7, PHP 5.3 doesn't work.

Even if we adjust the code to remove that "reason" (which we don't want to),
it may still fail for other reasons, like needing ircmaxell/password-compat
and it failing on certain 5.3 versions.
It's not worth to even think about it: PHP 5.4+ it is.