Do proper command arguments escaping
Until now it was possible to inject macros in arguments,
as well as used backquotes in bare arguments, which caused
shell commands to be executed.

This was not that bad, as we didn't intend for non-privileged
users to administrate the Nagios configuration anyway.

Making that safer, however, is a new direction for nagadmin,
potentially allowing anyone to (hopefully) safely submit
whatever arguments they want.